requirements.txt
crepe
librosa
numpy
matplotlib
music21
scipy
python-dotenv
google-generativeai